+++ app/src/main/java/com/galvezsh/paging3/presentation/model/CharacterModel.kt
@@ -1,7 +1,7 @@
 package com.galvezsh.paging3.presentation.model
 
 data class CharacterModel(
-    val id: Short,
+    val id: Int,
     val name: String,
     val isAlive: Boolean,
     val type: String,
